requirements.txt
langchain
langchain-google-genai
langchain_core
langgraph
google-generativeai
python-dotenv
langchain_openai
langchain-huggingface
numpy
scikit-learn